Moved Data.py to multiple files. Some other changes because I tried neovim
