+++ HqlCompiler/PolarsTools.py
@@ -1,7 +1,7 @@
 import polars as pl
-import logging
 
 class pltools():
+    @staticmethod
     def advance(columns:list[pl.DataFrame]) -> list[pl.DataFrame]:
         new = []
         name = columns[0].columns[0]
@@ -10,6 +10,7 @@ class pltools():
             
         return new
 
+    @staticmethod
     def merge(dfs:list[pl.DataFrame]):
         # Get counts for each column, knowing where we have conflicts.
         columns = {}
@@ -41,13 +42,15 @@ class pltools():
     # Fields is a list of the given path names.
     # host.name -> ['host', 'name']
     # Returns a df representation of that field, maintains nested-ness
-    def get_element(df:pl.DataFrame, path:list[str], index:int=0):
+    @staticmethod
+    def get_element(df:pl.DataFrame, path:list[str]):
         expr = pltools.path_to_expr(path)
         return df.select(expr)
 
     # Gets the value of an element, does not preserve df structure
     # Just returns the value
     # So for a base value, a series, and for a field that's a struct, a struct dataframe.
+    @staticmethod
     def get_element_value(df:pl.DataFrame, path:list[str]):
         expr = pltools.path_to_expr_value(path)
         data = df.select(expr)
@@ -57,6 +60,7 @@ class pltools():
         else:
             return data.to_series()
     
+    @staticmethod
     def build_element(name:list[str], data):
         if len(name) == 1:
             return pl.DataFrame({name[0]: data})
@@ -64,6 +68,7 @@ class pltools():
         new = pltools.build_element(name[1:], data)
         return pl.DataFrame({name[0]: new.to_struct()})
 
+    @staticmethod
     def path_to_expr_value(path:list[str]):
         # build selector
         expr = pl.col(path[0])
@@ -72,6 +77,7 @@ class pltools():
             
         return expr
 
+    @staticmethod
     def path_to_expr(path:list[str]):
         expr = pltools.path_to_expr_value(path)
         expr_str = 'pl.col(a).struct.field(c)'
