+++ HqlCompiler/Types/Elasticsearch.py
@@ -1,4 +1,4 @@
-from HqlCompiler.Types.Hql import HqlTypes as Hql
+from HqlCompiler.Types.Hql import HqlTypes as hqlt
 import logging
 
 from HqlCompiler.Exceptions import *
@@ -21,85 +21,85 @@ class ESTypes():
         return get_type(f'elasticsearch_{name}')
             
     @register_type('elasticsearch_scaled_float')
-    class scaled_float(ESType, Hql.decimal):
+    class scaled_float(ESType, hqlt.decimal):
         ...
 
     @register_type('elasticsearch_half_float')
-    class half_float(ESType, Hql.float):
+    class half_float(ESType, hqlt.float):
         ...
     
     @register_type('elasticsearch_float') 
-    class float(ESType, Hql.float):
+    class float(ESType, hqlt.float):
         ...
         
     @register_type('elasticsearch_double')
-    class double(ESType, Hql.double):
+    class double(ESType, hqlt.double):
         ...
     
     @register_type('elasticsearch_byte') 
-    class byte(ESType, Hql.byte):
+    class byte(ESType, hqlt.byte):
         ...
     
     @register_type('elasticsearch_short') 
-    class short(ESType, Hql.short):
+    class short(ESType, hqlt.short):
         ...
     
     @register_type('elasticsearch_integer') 
-    class integer(ESType, Hql.int):
+    class integer(ESType, hqlt.int):
         ...
     
     @register_type('elasticsearch_long') 
-    class long(ESType, Hql.long):
+    class long(ESType, hqlt.long):
         ...
     
     @register_type('elasticsearch_unsigned_long') 
-    class unsigned_long(ESType, Hql.ulong):
+    class unsigned_long(ESType, hqlt.ulong):
         ...
     
     @register_type('elasticsearch_ip') 
-    class ip(ESType, Hql.string):
+    class ip(ESType, hqlt.string):
         ...
         
     @register_type('elasticsearch_date')
-    class date(ESType, Hql.datetime):
+    class date(ESType, hqlt.datetime):
         ...
     
     @register_type('elasticsearch_date_nanos') 
-    class date_nanos(ESType, Hql.datetime):
+    class date_nanos(ESType, hqlt.datetime):
         ...
         
     @register_type('elasticsearch_date_range')
-    class date_range(ESType, Hql.range):
+    class date_range(ESType, hqlt.range):
         def __init__(self):
             ESTypes.ESType.__init__(self)
-            Hql.range.__init__(self, Hql.datetime)
+            hqlt.range.__init__(self, hqlt.datetime)
     
     @register_type('elasticsearch_integer_range') 
-    class integer_range(ESType, Hql.range):
+    class integer_range(ESType, hqlt.range):
         def __init__(self):
             ESTypes.ESType.__init__(self)
-            Hql.range.__init__(self, Hql.datetime)
+            hqlt.range.__init__(self, hqlt.datetime)
 
     @register_type('elasticsearch_float_range')
-    class float_range(ESType, Hql.range):
+    class float_range(ESType, hqlt.range):
         def __init__(self):
             ESTypes.ESType.__init__(self)
-            Hql.range.__init__(self, Hql.datetime)
+            hqlt.range.__init__(self, hqlt.datetime)
         
     @register_type('elasticsearch_long_range')   
-    class long_range(ESType, Hql.range):
+    class long_range(ESType, hqlt.range):
         def __init__(self):
             ESTypes.ESType.__init__(self)
-            Hql.range.__init__(self, Hql.datetime)
+            hqlt.range.__init__(self, hqlt.datetime)
         
     @register_type('elasticsearch_double_range') 
-    class double_range(ESType, Hql.range):
+    class double_range(ESType, hqlt.range):
         def __init__(self):
             ESTypes.ESType.__init__(self)
-            Hql.range.__init__(self, Hql.datetime)
+            hqlt.range.__init__(self, hqlt.datetime)
 
     @register_type('elasticsearch_ip_range') 
-    class ip_range(ESType, Hql.range):
+    class ip_range(ESType, hqlt.range):
         def __init__(self, ip_type:int):
             ESTypes.ESType.__init__(self)
             
@@ -107,64 +107,64 @@ class ESTypes():
                 raise CompilerException(f"Invalid IP version type {ip_type} passed to Elasticsearch type")
             
             if ip_type == 4:
-                Hql.range.__init__(self, type=Hql.ip4)
+                hqlt.range.__init__(self, type=hqlt.ip4)
             else:
-                Hql.range.__init__(self, type=Hql.ip6)
+                hqlt.range.__init__(self, type=hqlt.ip6)
     
     @register_type('elasticsearch_keyword') 
-    class keyword(ESType, Hql.string):
+    class keyword(ESType, hqlt.string):
         ...
     
     @register_type('elasticsearch_constant_keyword') 
-    class constant_keyword(ESType, Hql.string):
+    class constant_keyword(ESType, hqlt.string):
         ...
     
     @register_type('elasticsearch_wildcard') 
-    class wildcard(ESType, Hql.string):
+    class wildcard(ESType, hqlt.string):
         ...
     
     @register_type('elasticsearch_binary') 
-    class binary(ESType, Hql.string):
+    class binary(ESType, hqlt.string):
         ...
     
     @register_type('elasticsearch_object') 
-    class object(ESType, Hql.object):
+    class object(ESType, hqlt.object):
         ...
     
     @register_type('elasticsearch_flattened') 
-    class flattened(ESType, Hql.object):
+    class flattened(ESType, hqlt.object):
         ...
         
     @register_type('elasticsearch_nested')
-    class nested(ESType, Hql.object):
+    class nested(ESType, hqlt.object):
         ...
     
     @register_type('elasticsearch_alias') 
-    class alias(ESType, Hql.string):
+    class alias(ESType, hqlt.string):
         def __init__(self):
             logging.warn("Elasticsearch type 'alias' not implemented at the moment")
             ESTypes.ESType.__init__(self)
         
     @register_type('elasticsearch_point')
-    class point(ESType, Hql.multivalue):
+    class point(ESType, hqlt.multivalue):
         def __init__(self):
             ESTypes.ESType.__init__(self)
-            Hql.multivalue.__init__(self, Hql.float)
+            hqlt.multivalue.__init__(self, hqlt.float)
         
     @register_type('elasticsearch_geo_point')
-    class geo_point(ESType, Hql.multivalue):
+    class geo_point(ESType, hqlt.multivalue):
          def __init__(self):
             ESTypes.ESType.__init__(self)
-            Hql.multivalue.__init__(self, Hql.float)
+            hqlt.multivalue.__init__(self, hqlt.float)
     
     @register_type('elasticsearch_shape') 
-    class shape(ESType, Hql.matrix):
+    class shape(ESType, hqlt.matrix):
         def __init__(self):
             ESTypes.ESType.__init__(self)
-            Hql.matrix.__init__(self, Hql.float)
+            hqlt.matrix.__init__(self, hqlt.float)
                     
     @register_type('elasticsearch_geo_shape')   
-    class geo_shape(ESType, Hql.matrix):
+    class geo_shape(ESType, hqlt.matrix):
         def __init__(self):
             ESTypes.ESType.__init__(self)
-            Hql.matrix.__init__(self, Hql.float)
+            hqlt.matrix.__init__(self, hqlt.float)
